Fix linux issues

+++ source/Graphics/VulkanWrapper.cpp
@@ -1132,7 +1132,7 @@ void CmdBuildTLAS(TLAS& tlas, const std::vector<BLASInstance>& blasInstances) {
     barrier.dstAccessMask = VK_ACCESS_ACCELERATION_STRUCTURE_WRITE_BIT_KHR;
     vkCmdPipelineBarrier(cmd.buffer, VK_PIPELINE_STAGE_TRANSFER_BIT, VK_PIPELINE_STAGE_ACCELERATION_STRUCTURE_BUILD_BIT_KHR, 0, 1, &barrier, 0, nullptr, 0, nullptr);
 
-    VkAccelerationStructureBuildRangeInfoKHR buildOffsetInfo = {instances.size(), 0, 0, 0};
+    VkAccelerationStructureBuildRangeInfoKHR buildOffsetInfo = {uint32_t(instances.size()), 0, 0, 0};
     const VkAccelerationStructureBuildRangeInfoKHR* pBuildOffsetInfo = &buildOffsetInfo;
     _ctx.vkCmdBuildAccelerationStructuresKHR(cmd.buffer, 1, &res->buildInfo, &pBuildOffsetInfo);
     // todo: hash and only update mode if nothing changed
